Deduplicate artifact type registry instances
Mapping the attributes of an artifact set using the artifact type registry is expensive.

This commit allows us to take an artifact type regsitry and convert it into an interned immutable form,
allowing the cache itself to be used as a cache key.

This allows us to perform attribute mapping after the resolved variant cache instad of before it.

Further, this commit makes the resolved variant cache build-tree scoped instead of build scoped.